// dist/api/products/types.ts
/* eslint-disable */
// tslint:disable
/**
 * ECOMMERSE PROJECT MICROSERVICES
 * 
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface ProductsApi {
	imageUploadProduct: (request: Api.ImageUploadProductRequest.MultipartFormData | undefined) => Promise<ImageUploadProductResponse>
	updateProduct: (id: string, request: Api.CreateProduct | undefined) => Promise<UpdateProductResponse>
	createProduct: (request: Api.CreateProduct | undefined) => Promise<CreateProductResponse>
}

export type ImageUploadProductResponse = ImageUploadProduct200Response | ImageUploadProduct500Response

export interface ImageUploadProduct200Response {
	status: 200
	body: Api.ImageUploadProduct200Response
	headers?: never
}

export interface ImageUploadProduct500Response {
	status: 500
	body: Api.ImageUploadProduct500Response
	headers?: never
}

export type UpdateProductResponse = UpdateProduct200Response | UpdateProduct500Response

export interface UpdateProduct200Response {
	status: 200
	body: Api.UpdateProduct200Response
	headers?: never
}

export interface UpdateProduct500Response {
	status: 500
	body: Api.UpdateProduct500Response
	headers?: never
}

export type CreateProductResponse = CreateProduct201Response | CreateProduct500Response

export interface CreateProduct201Response {
	status: 201
	body: Api.CreateProduct201Response
	headers?: never
}

export interface CreateProduct500Response {
	status: 500
	body: Api.CreateProduct500Response
	headers?: never
}